==> 3 - Train Language Model Using FastAI.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

%cd /content/drive/MyDrive/Automate

!pip install wget pathos

!pip install nmslib

!pip install bcolz

!pip install pandas_summary

!pip install isoweek

!pip install sentencepiece!=0.1.90,!=0.1.91

%cd /content/drive/MyDrive/Automate

use_cache = False

from general_utils import get_step2_prerequisite_files



# Build Language Model From Docstrings

The goal is to build a language model using the docstrings, and use that language model to generate an embedding for each docstring.  

import torch,cv2
from lang_model_utils import lm_vocab, load_lm_vocab, train_lang_model
from general_utils import save_file_pickle, load_file_pickle
import logging
from pathlib import Path
from fastai.text.all import *
import pandas as pd

source_path = Path('./data/processed_data/')

with open(source_path/'train.docstring', 'r') as f:
    trn_data = f.readlines()

with open(source_path/'valid.docstring', 'r') as f:
    val_data = f.readlines()
    
with open(source_path/'test.docstring', 'r') as f:
    test_data = f.readlines()

df_data=trn_data+val_data
trn_raw=pd.DataFrame(trn_data)
val_raw=pd.DataFrame(val_data)
df_all = pd.concat([trn_raw, val_raw])

Preview what the raw data looks like: here are 10 docstrings

trn_raw[:10]

## Pre-process data for language model

We will use the class  `build_lm_vocab` to prepare our data for the language model

vocab = lm_vocab(max_vocab=50000,
                 min_freq=10)

# fit the transform on the training data, then transform
all_flat_idx = vocab.fit_transform_flattened(df_data)

Look at the transformed data

all_flat_idx[:10]

[vocab.itos[x] for x in all_flat_idx[:10]]

# apply transform to validation data
val_flat_idx = vocab.transform_flattened(val_data)

Save files for later use


vocab.save('./data/lang_model/vocab_v2.cls')
save_file_pickle('./data/lang_model/trn_flat_idx_list.pkl_v2', all_flat_idx)
save_file_pickle('./data/lang_model/val_flat_idx_list.pkl_v2', val_flat_idx)

## Train Fast.AI Language Model

This model will read in files that were created and train a [fast.ai](https://github.com/fastai/fastai/tree/master/fastai) language model.  This model learns to predict the next word in the sentence using fast.ai's implementation of [AWD LSTM](https://github.com/salesforce/awd-lstm-lm).  

The goal of training this model is to build a general purpose feature extractor for text that can be used in downstream models.  In this case, we will utilize this model to produce embeddings for function docstrings.

#Create data loaders
splits = [list(range_of(trn_raw)), list(range(len(val_raw), len(df_all)))]
tfms = [ attrgetter("text"),Tokenizer.from_df(0), Numericalize()]
dsets = Datasets(df_all, [tfms], splits=splits, dl_type=LMDataLoader)

bs,sl = 104,72
dls = dsets.dataloaders(bs=bs, seq_len=sl)

dls.show_batch(max_n=3)

# create lang model data
config = awd_lstm_lm_config.copy()
config.update({'input_p': 0.6, 'output_p': 0.4, 'weight_p': 0.5, 'embed_p': 0.1, 'hidden_p': 0.2})
model = get_language_model(AWD_LSTM, len(dls.vocab), config=config)

opt_func = partial(Adam, wd=0.1, eps=1e-7)
cbs = [MixedPrecision(), GradientClip(0.1)] + rnn_cbs(alpha=2, beta=1)

learn = Learner(dls, model, loss_func=CrossEntropyLossFlat(), opt_func=opt_func, cbs=cbs, metrics=[accuracy, Perplexity()])

learn.fit_one_cycle(10, 5e-3, moms=(0.8,0.7,0.8), div=10)

Save language model and learner

learn.save('lang_model_learner_v2.fai')
lang_model_new = learn.model.eval()
torch.save(lang_model_new, './data/lang_model/lang_model_gpu_v2.torch')
torch.save(lang_model_new.cpu(), './data/lang_model/lang_model_cpu_v2.torch')

# Load Model and Encode All Docstrings

Now that we have trained the language model, the next step is to use the language model to encode all of the docstrings into a vector. 

from lang_model_utils import load_lm_vocab
vocab = load_lm_vocab('./data/lang_model/vocab_v2.cls')
idx_docs = vocab.transform(trn_data + val_data, max_seq_len=30, padding=False)
lang_model = torch.load('./data/lang_model/lang_model_gpu_v2.torch', 
                        map_location=lambda storage, loc: storage)

lang_model.eval()

**Note:** the below code extracts embeddings for docstrings one docstring at a time, which is very inefficient.  Ideally, you want to extract embeddings in batch but account for the fact that you will have padding, etc. when extracting the hidden states.  For this tutorial, we only provide this minimal example, however you are welcome to improve upon this and sumbit a PR!

from tqdm import *

def list2arr(l):
    "Convert list into pytorch Variable."
    return torch.Tensor(np.expand_dims(np.array(l), -1)).long().cpu()

def make_prediction_from_list(model, l):
    """
    Encode a list of integers that represent a sequence of tokens.  The
    purpose is to encode a sentence or phrase.

    Parameters
    -----------
    model : fastai language model
    l : list
        list of integers, representing a sequence of tokens that you want to encode

    """
    arr = list2arr(l)# turn list into pytorch Variable with bs=1
    model.reset()  # language model is stateful, so you must reset upon each prediction
    hidden_states = model(arr)[-1][-1] # RNN Hidden Layer output is last output, and only need the last layer

    #return avg-pooling, max-pooling, and last hidden state
    return hidden_states.mean(0), hidden_states.max(0)[0], hidden_states[-1]


def get_embeddings(lm_model, list_list_int):
    """
    Vectorize a list of sequences List[List[int]] using a fast.ai language model.

    Paramters
    ---------
    lm_model : fastai language model
    list_list_int : List[List[int]]
        A list of sequences to encode

    Returns
    -------
    tuple: (avg, mean, last)
        A tuple that returns the average-pooling, max-pooling over time steps as well as the last time step.
    """
    n_rows = len(list_list_int)
    n_dim = lm_model[0].emb_sz
    avgarr = np.empty((n_rows, n_dim))
    maxarr = np.empty((n_rows, n_dim))
    lastarr = np.empty((n_rows, n_dim))

    for i in tqdm_notebook(range(len(list_list_int))):
        avg_, max_, last_ = make_prediction_from_list(lm_model, list_list_int[i])
        avgarr[i,:] = avg_.data.numpy()
        maxarr[i,:] = max_.data.numpy()
        lastarr[i,:] = last_.data.numpy()

    return avgarr, maxarr, lastarr

%%time
avg_hs, max_hs, last_hs = get_embeddings(lang_model, idx_docs)

### Do the same thing for the test set

idx_docs_test = vocab.transform(test_data, max_seq_len=30, padding=False)
avg_hs_test, max_hs_test, last_hs_test = get_embeddings(lang_model, idx_docs_test)

# Save Language Model Embeddings For Docstrings

savepath = Path('./data/lang_model/')
np.save(savepath/'avg_emb_dim500_v2.npy', avg_hs)
np.save(savepath/'max_emb_dim500_v2.npy', max_hs)
np.save(savepath/'last_emb_dim500_v2.npy', last_hs)

# save the test set embeddings also
np.save(savepath/'avg_emb_dim500_test_v2.npy', avg_hs_test)
np.save(savepath/'max_emb_dim500_test_v2.npy', max_hs_test)
np.save(savepath/'last_emb_dim500_test_v2.npy', last_hs_test)

### Create search index using `nmslib` 

[nmslib](https://github.com/nmslib/nmslib) is a great library for doing nearest neighbor lookups, which we will use as a search engine for finding nearest neighbors of comments in vector-space.  

The convenience function `create_nmslib_search_index` builds this search index given a matrix of vectors as input.



from general_utils import create_nmslib_search_index
import nmslib
from lang_model_utils import Query2Emb
from pathlib import Path
import numpy as np
from lang_model_utils import load_lm_vocab
import torch

# Load matrix of vectors
loadpath = Path('./data/lang_model/')
avg_emb_dim500 = np.load(loadpath/'avg_emb_dim500_test_v2.npy')

# Build search index (takes about an hour on a p3.8xlarge)
dim500_avg_searchindex = create_nmslib_search_index(avg_emb_dim500)

# save search index
dim500_avg_searchindex.saveIndex('./data/lang_model/dim500_avg_searchindex.nmslib')

After building this search index with nmslib, we can do fast nearest-neighbor lookups.  We use the `Query2Emb` object to help convert strings to the embeddings: 

dim500_avg_searchindex = nmslib.init(method='hnsw', space='cosinesimil')
dim500_avg_searchindex.loadIndex('./data/lang_model/dim500_avg_searchindex.nmslib')

lang_model = torch.load('./data/lang_model/lang_model_cpu_v2.torch',map_location=torch.device('cpu'))
vocab = load_lm_vocab('./data/lang_model/vocab_v2.cls')

q2emb = Query2Emb(lang_model = lang_model.cpu(),
                  vocab = vocab)

The method `Query2Emb.emb_mean` will allow us to use the langauge model we trained earlier to generate a sentence embedding given a string.   Here is an example, `emb_mean` will return a numpy array of size (1, 500).

query = q2emb.emb_mean('Read data into pandas dataframe')
query.shape

**Make search engine to inspect semantic similarity of phrases**.  This will take 3 inputs:

1. `nmslib_index` - this is the search index we built above.  This object takes a vector and will return the index of the closest vector(s) according to cosine distance.  
2. `ref_data` - this is the data for which the index refer to, in this case will be the docstrings. 
3. `query2emb_func` - this is a function that will convert a string into an embedding.

class search_engine:
    def __init__(self, 
                 nmslib_index, 
                 ref_data, 
                 query2emb_func):
        
        self.search_index = nmslib_index
        self.data = ref_data
        self.query2emb_func = query2emb_func
    
    def search(self, str_search, k=3):
        query = self.query2emb_func(str_search)
        idxs, dists = self.search_index.knnQuery(query, k=k)
        
        for idx, dist in zip(idxs, dists):
            print(f'cosine dist:{dist:.4f}\n---------------\n', self.data[idx])

se = search_engine(nmslib_index=dim500_avg_searchindex,
                   ref_data = test_data,
                   query2emb_func = q2emb.emb_mean)

## Manually Inspect Phrase Similarity

Compare a user-supplied query vs. vectorized docstrings on test set.  We can see that similar phrases are not exactly the same, but the nearest neighbors are reasonable.  

import logging
logging.getLogger().setLevel(logging.ERROR)

se.search('read csv into pandas dataframe')

se.search('django files upload')

se.search('read data from source and target files and put into buckets .')

se.search('validate data using model level methods .')

se.search('send out email notification')

se.search('save pickle file')

se.search('rest api')
